1.0 Calculate depth of a full Binary tree from Preorder to 1.0 Find the Maximum Depth or Height of given Binary Tree
Given a binary tree, the task is to find the height of the tree. The height of the tree is the number of vertices in the tree from the root to the deepest node. 

Note: The height of an empty tree is 0 and the height of a tree with single node is 1.